ocr-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
# PaddleOCR 2.x requires NumPy 1.x; keep <2 for resolver compatibility with paddleocr<3.
numpy>=1.26,<2
Pillow>=10.3.0
# Contrib + headless: satisfies PaddleOCR 2.x (opencv-contrib) without GUI stack; install before paddleocr.
opencv-contrib-python-headless>=4.10.0.84
paddlepaddle>=2.6.1
# PaddleOCR 3.x re-exports the full PaddleX stack on import (breaks many slim Docker images). Stay on 2.x API (.ocr).
paddleocr>=2.7.0.3,<3.0.0
llama-cpp-python>=0.2.82
